+++ src/AppInstallerCLICore/Workflows/SourceFlow.cpp
@@ -272,6 +272,40 @@ namespace AppInstaller::CLI::Workflow
         }
     }
 
+    void EditSources(Execution::Context& context)
+    {
+        // We are assuming there is only one match, as SourceName is a required parameter.
+        const std::vector<Repository::SourceDetails>& sources = context.Get<Data::SourceList>();
+
+        for (const auto& sd : sources)
+        {
+            // Get the current source with this name.
+            Repository::Source targetSource{ sd.Name };
+            auto oldExplicitValue = sd.Explicit;
+
+            std::optional<bool> isExplicit;
+            if (context.Args.Contains(Execution::Args::Type::SourceEditExplicit))
+            {
+                isExplicit = Utility::TryConvertStringToBool(context.Args.GetArg(Execution::Args::Type::SourceEditExplicit));
+            }
+
+            Repository::SourceEdit edits{ isExplicit };
+            if (!targetSource.RequiresChanges(edits))
+            {
+                context.Reporter.Info() << Resource::String::SourceEditNoChanges(Utility::LocIndView{ sd.Name }) << std::endl;
+                continue;
+            }
+
+            context.Reporter.Info() << Resource::String::SourceEditOne(Utility::LocIndView{ sd.Name }) << std::endl;
+            targetSource.Edit(edits);
+
+            // Output updated source information. Since only Explicit is editable, we will only list that field. The name of the source being edited is listed prior to the edits.
+            Execution::TableOutput<3> table(context.Reporter, { Resource::String::SourceListField, Resource::String::SourceEditOldValue, Resource::String::SourceEditNewValue });
+            table.OutputLine({ Resource::LocString(Resource::String::SourceListExplicit), std::string{ Utility::ConvertBoolToString(oldExplicitValue) }, std::string{ Utility::ConvertBoolToString(isExplicit.value()) } });
+            table.Complete();
+        }
+    }
+
     void QueryUserForSourceReset(Execution::Context& context)
     {
         if (!context.Args.Contains(Execution::Args::Type::ForceSourceReset))