+++ src/AppInstallerCommonCore/Public/winget/HttpClientHelper.h
@@ -2,8 +2,10 @@
 // Licensed under the MIT License.
 #pragma once
 #include <winget/Certificates.h>
+#include <winget/SharedThreadGlobals.h>
 #include <cpprest/http_client.h>
 #include <cpprest/json.h>
+#include <memory>
 #include <optional>
 #include <vector>
 
@@ -34,7 +36,7 @@ namespace AppInstaller::Http
 
         std::optional<web::json::value> HandleGet(const utility::string_t& uri, const HttpRequestHeaders& headers = {}, const HttpRequestHeaders& authHeaders = {}, const HttpResponseHandler& customHandler = {}) const;
 
-        void SetPinningConfiguration(const Certificates::PinningConfiguration& configuration);
+        void SetPinningConfiguration(const Certificates::PinningConfiguration& configuration, std::shared_ptr<ThreadLocalStorage::ThreadGlobals> threadGlobals = {});
 
     protected:
         std::optional<web::json::value> ValidateAndExtractResponse(const web::http::http_response& response) const;