+++ src/Microsoft.Management.Deployment/AddPackageCatalogOptions.h
@@ -3,6 +3,7 @@
 #pragma once
 #include "AddPackageCatalogOptions.g.h"
 #include "public/ComClsids.h"
+#include <winget/ModuleCountBase.h>
 
 namespace winrt::Microsoft::Management::Deployment::implementation
 {
@@ -44,7 +45,7 @@ namespace winrt::Microsoft::Management::Deployment::implementation
 #if !defined(INCLUDE_ONLY_INTERFACE_METHODS)
 namespace winrt::Microsoft::Management::Deployment::factory_implementation
 {
-    struct AddPackageCatalogOptions : AddPackageCatalogOptionsT<AddPackageCatalogOptions, implementation::AddPackageCatalogOptions>
+    struct AddPackageCatalogOptions : AddPackageCatalogOptionsT<AddPackageCatalogOptions, implementation::AddPackageCatalogOptions>, AppInstaller::WinRT::ModuleCountBase
     {
     };
 }