+++ src/Microsoft.Management.Deployment/ComClsids.cpp
@@ -17,6 +17,7 @@
 #include "RepairOptions.h"
 #include "AddPackageCatalogOptions.h"
 #include "RemovePackageCatalogOptions.h"
+#include "EditPackageCatalogOptions.h"
 #pragma warning( pop )
 
 namespace winrt::Microsoft::Management::Deployment
@@ -70,6 +71,10 @@ namespace winrt::Microsoft::Management::Deployment
         else if (IsEqualCLSID(clsid, WINGET_INPROC_COM_CLSID_RemovePackageCatalogOptions))
         {
             return __uuidof(winrt::Microsoft::Management::Deployment::implementation::RemovePackageCatalogOptions);
+        }
+        else if (IsEqualCLSID(clsid, WINGET_INPROC_COM_CLSID_EditPackageCatalogOptions))
+        {
+            return __uuidof(winrt::Microsoft::Management::Deployment::implementation::EditPackageCatalogOptions);
         }
         else
         {